+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Features
 
+- Add new `--no-ignore-parent` flag, see #787 (@will459)
+
 ## Bugfixes
 
 - Set default path separator to `/` in MSYS, see #537 and #730 (@aswild)